In CelestialObjectMoving and PlanetMoving add a passpassParameter function

+++ Assets/Script/Stars/Moving/CelestialObjectMoving.cs
@@ -6,7 +6,7 @@ public class CelestialObjectMoving: MonoBehaviour
 {
     protected float rotationSpeed;
 
-    public virtual void passParameter(int rotationSpeed) {
+    public virtual void passParameter(float rotationSpeed) {
         this.rotationSpeed = rotationSpeed;
     }
